pyproject.toml
[project]
name = "lite-llm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.73.0",
    "python-dotenv>=1.1.1",
]
